requirements.txt
streamlit
pandas
nltk
scikit-learn
wordcloud
openai
python-dotenv
